+++ R/globals.R
@@ -1,6 +1,6 @@
 utils::globalVariables(c("RMSE", "adjR", "control", "controlID",
                          "control_coefs", "estimate", "estimate_FE", "feols",
-                         "fixed", "index", "is_fe", "lower", "p",
+                         "fixed", "included", "index", "is_fe", "lower", "p",
                          "ribbon.group", "se", "se_type", "sig", "sig.level",
                          "statistic", "term", "upper", "value", "."))